fire optymalizacja reloadu danych po dodaniu do listy zakupów
